Init package
This package might use to genarate global ID for microservices system or can be use in other system need unique integer ID(make a shortlink for ex).

The ID has 64 bits in total, 41 bit for time from epoch time until now, 12 bits for sequence number in milisecond(it mean 1 milisecond we can genarate
2^12 number), 11 bits for shard.

Current ly epoch time is "time.Parse("02/01/2006", "22/06/2022")", in the future we might support to change the epoch time.

Currently the code not cover to get sequence number each milisecond(and make sure it unique), in future we might do this by use mutex.Lock or use atomic package in Go. Note that if in one milisecond
we have sequence number larger than 2^12 then we need to wait until next milisecond.